requirements.txt
python-stdnum==2.1